graphはqulacs-osakaに行ったので消した　create_time_gate preへ
