+++ skqulacs/circuit/__init__.py
@@ -6,4 +6,3 @@ from .pre_defined import (
     create_deepqsv,
     create_farhi_watle,
 )
-from .graph import show_blochsphere
